CI: Move the whole repo to a Makefile (#249)
* CI: Move the whole repo to a Makefile

#### Problem

The memo repo is still using the zx scripts for everything, even though
other repos are now on Makefiles.

#### Summary of changes

Move everything over to the Makefile! As part of this, I had to rename
the `p-memo` directory to `pinocchio`, because otherwise the Makefile
misinterprets the target name as `p/memo`. We could also move the
programs into a common subdirectory, but this seemed like an easier
option.

* Fix various failures

* update deps for audit
* add format job to js-legacy
* add allowed warnings in clippy
* add more words to dictionary
* wait longer for js tests